requirements.txt
django>=6.0.3
django-redis>=6.0.0
django-scope>=1.0.1
djangorestframework>=3.17.1
daphne>=4.2.1
channels>=4.3.2
channels-redis>=4.3.0
psycopg[binary]>=3.3.3
pillow>=12.2.0
python-escpos>=3.1
pyusb>=1.3.1
aiogram>=3.27.0
requests>=2.33.0
python-dotenv>=1.2.2
redis>=7.4.0
celery>=5.5.0
